Added InstanceProps<T> and resolvePropertyPath to exports of core (undecided)
